Bugfixes, alterations for Acrobat XI environment
What a nightmare, by the way, the try/catch statement isn't working
to prevent the exception from bubbling up and ruining everything in
AcroJS's compiler. The script is still not functional - two big issues
remaining, and that's that util.stringFromStream doesn't handle unicode
input (for the love of god, why?), and the other one, which is more
under my control, is that I think the print loop removes all <<>> and []
characters after one iteration, bringing the program to a crashing halt
when self.getTextBoxes() returns undefined. Because there aren't any
text boxes. Should be easy enough to fix, but that's a job for tomorrow
me.